pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predictive-ling-cli"
version = "0.2.0"
description = "Predictive Linguistics CLI - No API keys required"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "gtts>=2.4.0",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "ruff>=0.1.0", "mypy>=1.0.0"]

[project.scripts]
predictive-ling = "predictive_ling_cli.cli:main"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
